+++ angular-workspace/projects/ni/nimble-angular/src/directives/table-column/text/tests/nimble-table-column-text.directive.spec.ts
@@ -21,7 +21,14 @@ describe('NimbleTableColumnText', () => {
         @Component({
             template: `
                 <nimble-table>
-                    <nimble-table-column-text #column field-name="field1" placeholder="no value"></nimble-table-column-text>
+                    <nimble-table-column-text
+                        #column
+                        column-id="my-column"
+                        field-name="field1"
+                        placeholder="no value"
+                        action-menu-slot="my-slot"
+                        action-menu-label="my menu"
+                    ></nimble-table-column-text>
                 </nimble-table>
             `
         })
@@ -50,9 +57,24 @@ describe('NimbleTableColumnText', () => {
             expect(nativeElement.fieldName).toBe('field1');
         });
 
-        it('cwill use template string values for placeholder', () => {
+        it('will use template string values for placeholder', () => {
             expect(directive.placeholder).toBe('no value');
             expect(nativeElement.placeholder).toBe('no value');
+        });
+
+        it('will use template string values for actionMenuSlot', () => {
+            expect(directive.actionMenuSlot).toBe('my-slot');
+            expect(nativeElement.actionMenuSlot).toBe('my-slot');
+        });
+
+        it('will use template string values for actionMenuLabel', () => {
+            expect(directive.actionMenuLabel).toBe('my menu');
+            expect(nativeElement.actionMenuLabel).toBe('my menu');
+        });
+
+        it('will use template string values for columnId', () => {
+            expect(directive.columnId).toBe('my-column');
+            expect(nativeElement.columnId).toBe('my-column');
         });
     });
 
@@ -60,7 +82,14 @@ describe('NimbleTableColumnText', () => {
         @Component({
             template: `
                 <nimble-table>
-                    <nimble-table-column-text #column [field-name]="field" [placeholder]="placeholder"></nimble-table-column-text>
+                    <nimble-table-column-text
+                        #column
+                        [column-id]="columnId"
+                        [field-name]="field"
+                        [placeholder]="placeholder"
+                        [actionMenuSlot]="actionMenuSlot"
+                        [actionMenuLabel]="actionMenuLabel"
+                    ></nimble-table-column-text>
                 </nimble-table>
             `
         })
@@ -69,6 +98,9 @@ describe('NimbleTableColumnText', () => {
             @ViewChild('column', { read: ElementRef }) public elementRef: ElementRef<TableColumnText>;
             public field = 'field1';
             public placeholder = 'no value';
+            public actionMenuSlot = 'my-slot';
+            public actionMenuLabel = 'my menu';
+            public columnId = 'my-column';
         }
 
         let fixture: ComponentFixture<TestHostComponent>;
@@ -106,6 +138,39 @@ describe('NimbleTableColumnText', () => {
 
             expect(directive.placeholder).toBe('foo');
             expect(nativeElement.placeholder).toBe('foo');
+        });
+
+        it('can be configured with property binding for actionMenuSlot', () => {
+            expect(directive.actionMenuSlot).toBe('my-slot');
+            expect(nativeElement.actionMenuSlot).toBe('my-slot');
+
+            fixture.componentInstance.actionMenuSlot = 'new-slot';
+            fixture.detectChanges();
+
+            expect(directive.actionMenuSlot).toBe('new-slot');
+            expect(nativeElement.actionMenuSlot).toBe('new-slot');
+        });
+
+        it('can be configured with property binding for actionMenuLabel', () => {
+            expect(directive.actionMenuLabel).toBe('my menu');
+            expect(nativeElement.actionMenuLabel).toBe('my menu');
+
+            fixture.componentInstance.actionMenuLabel = 'another menu';
+            fixture.detectChanges();
+
+            expect(directive.actionMenuLabel).toBe('another menu');
+            expect(nativeElement.actionMenuLabel).toBe('another menu');
+        });
+
+        it('can be configured with property binding for columnId', () => {
+            expect(directive.columnId).toBe('my-column');
+            expect(nativeElement.columnId).toBe('my-column');
+
+            fixture.componentInstance.columnId = 'new-column';
+            fixture.detectChanges();
+
+            expect(directive.columnId).toBe('new-column');
+            expect(nativeElement.columnId).toBe('new-column');
         });
     });
 
@@ -113,7 +178,14 @@ describe('NimbleTableColumnText', () => {
         @Component({
             template: `
                 <nimble-table>
-                    <nimble-table-column-text #column [attr.field-name]="field" [attr.placeholder]="placeholder"></nimble-table-column-text>
+                    <nimble-table-column-text
+                        #column
+                        [attr.column-id]="columnId"
+                        [attr.field-name]="field"
+                        [attr.placeholder]="placeholder"
+                        [attr.action-menu-slot]="actionMenuSlot"
+                        [attr.action-menu-label]="actionMenuLabel"
+                    ></nimble-table-column-text>
                 </nimble-table>
             `
         })
@@ -122,6 +194,9 @@ describe('NimbleTableColumnText', () => {
             @ViewChild('column', { read: ElementRef }) public elementRef: ElementRef<TableColumnText>;
             public field = 'field1';
             public placeholder = 'no value';
+            public actionMenuSlot = 'my-slot';
+            public actionMenuLabel = 'my menu';
+            public columnId = 'my-column';
         }
 
         let fixture: ComponentFixture<TestHostComponent>;
@@ -139,7 +214,7 @@ describe('NimbleTableColumnText', () => {
             nativeElement = fixture.componentInstance.elementRef.nativeElement;
         });
 
-        it('can be configured with property binding for fieldName', () => {
+        it('can be configured with attribute binding for fieldName', () => {
             expect(directive.fieldName).toBe('field1');
             expect(nativeElement.fieldName).toBe('field1');
 
@@ -150,7 +225,7 @@ describe('NimbleTableColumnText', () => {
             expect(nativeElement.fieldName).toBe('field2');
         });
 
-        it('can be configured with property binding for placeholder', () => {
+        it('can be configured with attribute binding for placeholder', () => {
             expect(directive.placeholder).toBe('no value');
             expect(nativeElement.placeholder).toBe('no value');
 
@@ -159,6 +234,39 @@ describe('NimbleTableColumnText', () => {
 
             expect(directive.placeholder).toBe('foo');
             expect(nativeElement.placeholder).toBe('foo');
+        });
+
+        it('can be configured with attribute binding for actionMenuSlot', () => {
+            expect(directive.actionMenuSlot).toBe('my-slot');
+            expect(nativeElement.actionMenuSlot).toBe('my-slot');
+
+            fixture.componentInstance.actionMenuSlot = 'new-slot';
+            fixture.detectChanges();
+
+            expect(directive.actionMenuSlot).toBe('new-slot');
+            expect(nativeElement.actionMenuSlot).toBe('new-slot');
+        });
+
+        it('can be configured with attribute binding for actionMenuLabel', () => {
+            expect(directive.actionMenuLabel).toBe('my menu');
+            expect(nativeElement.actionMenuLabel).toBe('my menu');
+
+            fixture.componentInstance.actionMenuLabel = 'another menu';
+            fixture.detectChanges();
+
+            expect(directive.actionMenuLabel).toBe('another menu');
+            expect(nativeElement.actionMenuLabel).toBe('another menu');
+        });
+
+        it('can be configured with attribute binding for columnId', () => {
+            expect(directive.columnId).toBe('my-column');
+            expect(nativeElement.columnId).toBe('my-column');
+
+            fixture.componentInstance.columnId = 'new-column';
+            fixture.detectChanges();
+
+            expect(directive.columnId).toBe('new-column');
+            expect(nativeElement.columnId).toBe('new-column');
         });
     });
 });