+++ packages/nimble-components/src/table/index.ts
@@ -1,4 +1,9 @@
-import { attr, observable } from '@microsoft/fast-element';
+import {
+    attr,
+    Observable,
+    observable,
+    Notifier
+} from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
     ColumnDef as TanStackColumnDef,
@@ -9,11 +14,16 @@ import {
     getCoreRowModel as tanStackGetCoreRowModel,
     TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
-import type { TableColumn } from '../table-column/base';
+import { TableColumn } from '../table-column/base';
 import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableRecord, TableRowState, TableValidity } from './types';
+import type {
+    TableActionMenuToggleEventDetail,
+    TableRecord,
+    TableRowState,
+    TableValidity
+} from './types';
 import { Virtualizer } from './models/virtualizer';
 
 declare global {
@@ -37,8 +47,29 @@ export class Table<
     @observable
     public tableData: TableRowState<TData>[] = [];
 
+    /**
+     * @internal
+     */
     @observable
-    public readonly columns: TableColumn[] = [];
+    public columns: TableColumn[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
+    public readonly childItems: Element[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
+    public actionMenuSlots: string[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
+    public openActionMenuRecordId?: string;
 
     /**
      * @internal
@@ -63,6 +94,7 @@ export class Table<
     private readonly table: TanStackTable<TData>;
     private options: TanStackTableOptionsResolved<TData>;
     private readonly tableValidator = new TableValidator();
+    private columnNotifiers: Notifier[] = [];
 
     public constructor() {
         super();
@@ -84,7 +116,55 @@ export class Table<
         this.setTableData(newData);
     }
 
-    public idFieldNameChanged(
+    public override connectedCallback(): void {
+        super.connectedCallback();
+        this.virtualizer.connectedCallback();
+        this.validateAndObserveColumns();
+    }
+
+    public override disconnectedCallback(): void {
+        super.disconnectedCallback();
+        this.virtualizer.disconnectedCallback();
+        this.removeColumnObservers();
+    }
+
+    public checkValidity(): boolean {
+        return this.tableValidator.isValid();
+    }
+
+    /**
+     * @internal
+     *
+     * The event handler that is called when a notifier detects a change. Notifiers are added
+     * to each column, so `source` is expected to be an instance of `TableColumn`, and `args`
+     * is the string name of the property that changed on that column.
+     */
+    public handleChange(source: unknown, args: unknown): void {
+        if (source instanceof TableColumn) {
+            if (args === 'columnId') {
+                this.validateColumnIds();
+            }
+        }
+    }
+
+    public onRowActionMenuBeforeToggle(
+        event: CustomEvent<TableActionMenuToggleEventDetail>
+    ): void {
+        this.openActionMenuRecordId = event.detail.recordIds[0];
+        this.$emit('action-menu-beforetoggle', event.detail);
+    }
+
+    public onRowActionMenuToggle(
+        event: CustomEvent<TableActionMenuToggleEventDetail>
+    ): void {
+        this.$emit('action-menu-toggle', event.detail);
+    }
+
+    protected childItemsChanged(): void {
+        void this.updateColumnsFromChildItems();
+    }
+
+    protected idFieldNameChanged(
         _prev: string | undefined,
         _next: string | undefined
     ): void {
@@ -93,17 +173,59 @@ export class Table<
         this.setTableData(this.table.options.data);
     }
 
-    public override connectedCallback(): void {
-        super.connectedCallback();
-        this.virtualizer.connectedCallback();
+    protected columnsChanged(
+        _prev: TableColumn[] | undefined,
+        _next: TableColumn[]
+    ): void {
+        if (!this.$fastController.isConnected) {
+            return;
+        }
+
+        this.validateAndObserveColumns();
+
+        const slots = new Set<string>();
+        for (const column of this.columns) {
+            if (column.actionMenuSlot) {
+                slots.add(column.actionMenuSlot);
+            }
+        }
+        this.actionMenuSlots = Array.from(slots);
     }
 
-    public override disconnectedCallback(): void {
-        this.virtualizer.disconnectedCallback();
+    private removeColumnObservers(): void {
+        this.columnNotifiers.forEach(notifier => {
+            notifier.unsubscribe(this);
+        });
+        this.columnNotifiers = [];
     }
 
-    public checkValidity(): boolean {
-        return this.tableValidator.isValid();
+    private validateAndObserveColumns(): void {
+        this.removeColumnObservers();
+
+        for (const column of this.columns) {
+            const notifier = Observable.getNotifier(column);
+            notifier.subscribe(this);
+            this.columnNotifiers.push(notifier);
+        }
+
+        this.validateColumnIds();
+    }
+
+    private validateColumnIds(): void {
+        this.tableValidator.validateColumnIds(
+            this.columns.map(x => x.columnId)
+        );
+        this.canRenderRows = this.checkValidity();
+    }
+
+    private async updateColumnsFromChildItems(): Promise<void> {
+        const definedElements = this.childItems.map(async item => (item.matches(':not(:defined)')
+            ? customElements.whenDefined(item.localName)
+            : Promise.resolve()));
+        await Promise.all(definedElements);
+        this.columns = this.childItems.filter(
+            (x): x is TableColumn => x instanceof TableColumn
+        );
     }
 
     private setTableData(newData: readonly TData[]): void {