+++ packages/nimble-tokens/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-tokens
 
-This log was last generated on Fri, 10 Feb 2023 19:40:36 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Feb 2023 16:15:37 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 4.7.0
+
+Tue, 14 Feb 2023 16:15:37 GMT
+
+### Minor changes
+
+- Add font asset for Noto Serif ([ni/nimble@d9e4787](https://github.com/ni/nimble/commit/d9e478746435367f4909d7592d859747f4b67145))
 
 ## 4.6.0